Adds serving labels to recipe cells.
